refactor(actions): Refactoring of some classes in order to tidy up the ccw.editors.clojure Java namespace.
