+++ ccw.core/src/java/ccw/editors/clojure/actions/GotoNextMemberAction.java
@@ -8,25 +8,28 @@
  * Contributors: 
  *    Laurent PETIT - initial API and implementation
  *******************************************************************************/
-package ccw.editors.clojure;
+package ccw.editors.clojure.actions;
 
 import org.eclipse.core.runtime.Assert;
 import org.eclipse.jface.action.Action;
 
-public class SelectTopLevelSExpressionAction extends Action {
+import ccw.editors.clojure.ClojureEditor;
+import ccw.editors.clojure.ClojureEditorMessages;
 
-	public final static String ID = "SelectTopLevelSExpressionAction"; //$NON-NLS-1$
+public class GotoNextMemberAction extends Action {
+
+	public final static String ID = "ClojureGotoNextMember"; //$NON-NLS-1$
 
 	private final ClojureEditor editor;
 
-	public SelectTopLevelSExpressionAction(ClojureEditor editor) {
-		super(ClojureEditorMessages.SelectTopLevelSExpressionAction_label);
+	public GotoNextMemberAction(ClojureEditor editor) {
+		super(ClojureEditorMessages.GotoNextMemberAction_label);
 		Assert.isNotNull(editor);
 		this.editor= editor;
 		setEnabled(true);
 	}
 
 	public void run() {
-		editor.selectTopLevelSExpression();
+		editor.gotoEndOfMember();
 	}
 }
