+++ ccw.core/src/java/ccw/editors/clojure/actions/GotoPreviousMemberAction.java
@@ -8,10 +8,13 @@
  * Contributors: 
  *    Laurent PETIT - initial API and implementation
  *******************************************************************************/
-package ccw.editors.clojure;
+package ccw.editors.clojure.actions;
 
 import org.eclipse.core.runtime.Assert;
 import org.eclipse.jface.action.Action;
+
+import ccw.editors.clojure.ClojureEditor;
+import ccw.editors.clojure.ClojureEditorMessages;
 
 public class GotoPreviousMemberAction extends Action {
