+++ ccw.core/src/java/ccw/editors/clojure/actions/SwitchNamespaceAction.java
@@ -1,9 +1,12 @@
-package ccw.editors.clojure;
+package ccw.editors.clojure.actions;
 
 import org.eclipse.core.runtime.Assert;
 import org.eclipse.jface.action.Action;
 
 import ccw.CCWPlugin;
+import ccw.editors.clojure.ClojureEditor;
+import ccw.editors.clojure.ClojureEditorMessages;
+import ccw.editors.clojure.EvaluateTextUtil;
 import ccw.repl.Actions;
 import ccw.repl.REPLView;
